backend/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
web3==5.23.1
Pillow==8.3.2